feat: Have HTTP 403 check for SSO failure
AWS API Gateway will return a 403 for a present
but invalid (eg. expired) jwt token. Because
the UI doesn't handle these the same way it handles
the 401s the login screen is not loaded, forcing a user to
explictly go to the /auth/logout to clear the cookie, then go
to /login to get the login button to show up